requirements.txt
tensorflow==2.19.*
keras==3.6.*
tf-keras==2.19.*
h5py
pillow
numpy
matplotlib
streamlit
